+++ hata/discord/embed/embed_base.py
@@ -3,10 +3,11 @@ __all__ = (
     'EmbedThumbnail', 'EmbedVideo'
 )
 
+import warnings
 from datetime import datetime
 
 from ..color import Color
-from ..utils import url_cutter
+from ..utils import is_url, url_cutter
 
 
 EXTRA_EMBED_TYPES = frozenset(('application_news', 'article', 'gifv', 'image', 'link', 'tweet', 'video'))
@@ -703,6 +704,50 @@ class EmbedProvider:
         return self
 
 
+def _warn_bad_embed_author_constructor_parameter_order(name, icon_url):
+    """
+    Checks whether the embed author `name` and `icon_url`-s are passed the other way around. If not drops a warning.
+    
+    Parameters
+    ----------
+    name : `None`, `str`
+        Embed author name.
+    icon_url : `None`, `str`
+        Embed author icon url.
+    
+    Returns
+    -------
+    name : `None`, `str`
+        Embed author name.
+    icon_url : `None`, `str`
+        Embed author icon url.
+    """
+    if name is None:
+        if (icon_url is not None) and (not is_url(icon_url)):
+            switch = True
+        else:
+            switch = False
+    
+    else:
+        if is_url(name) and ((icon_url is None) or (not is_url(icon_url))):
+            switch = True
+        else:
+            switch = False
+    
+    if switch:
+        warnings.warn(
+            FutureWarning,
+            (
+                f'Embed author parameter order changed to `name, icon_url, url` from `icon_url`, `name`, `url` one. '
+                f'Using the old version is deprecated and will be removed in 2022 Jul.'
+            ),
+        )
+        
+        icon_url, name = name, icon_url
+    
+    return name, icon_url
+
+
 class EmbedAuthor:
     """
     Represents an embed's author.
@@ -720,19 +765,21 @@ class EmbedAuthor:
     """
     __slots__ = ('icon_url', 'name', 'proxy_icon_url', 'url')
     
-    def __init__(self, icon_url=None, name=None, url=None):
+    def __init__(self, name=None, icon_url=None, url=None):
         """
         Creates an embed author with the given parameters.
         
         Parameters
         ----------
-        icon_url : `None`, `str` = `None`, Optional
-            An url of the author's icon. Can be http(s) or attachment.
         name : `None`, `str` = `None`, Optional
             The name of the author.
+        icon_url : `None`, `str` = `None`, Optional
+            An url of the author's icon. Can be http(s) or attachment.
         url : `None`, `str` = `None`, Optional
             The url of the author.
         """
+        name, icon_url = _warn_bad_embed_author_constructor_parameter_order(name, icon_url)
+        
         self.icon_url = icon_url
         self.name = name
         self.proxy_icon_url = None